Split contents logic to separate file
